+++ bin/build.sh
@@ -1,0 +1,13 @@
+#!/usr/bin/env sh
+
+set -e
+
+bin="`dirname "$0"`"
+root="$bin/.."
+
+name="`basename "$(realpath "$root")"`"
+user="${FLUX_PUBLISH_DOCKER_USER:=fluxfw}"
+image="$user/$name"
+tag="`get-release-tag "$root"`"
+
+docker build "$root" --pull -t "$image:$tag" -t "$image:latest"